+++ BrowserDetection.php
@@ -11,6 +11,8 @@ namespace Piwik\Plugins\TrackingSpamPrevention;
 use Matomo\Network\IP;
 use Piwik\Cache as PiwikCache;
 use Piwik\Common;
+use Piwik\Container\StaticContainer;
+use Piwik\DeviceDetector\DeviceDetectorFactory;
 use Piwik\Option;
 use Piwik\Piwik;
 use Piwik\Plugins\TrackingSpamPrevention\BlockedIpRanges\IpRangeProviderInterface;
@@ -28,7 +30,13 @@ class BrowserDetection
         $browsers = [
             'HeadlessChrome',
             'PhantomJS',
-            'Electron'
+            'Electron',
+            'ApacheBench', // Load test
+            'Siege/', // Load test
+            'https://k6.io/', // Load test, Added full URL to handle case where UA can be fook6bar or fook6/
+            'Radview', // Load test
+            'Locust', // Load test
+            'Cypress' //Testing Tool
         ];
         foreach ($browsers as $browser) {
             if (stripos($userAgent, $browser) !== false) {
@@ -36,6 +44,16 @@ class BrowserDetection
             }
         }
         return false;
+    }
+
+    public function isLibrary($userAgent)
+    {
+        if (empty($userAgent)) {
+            return false;
+        }
+        $staticContainer = StaticContainer::get(DeviceDetectorFactory::class)->makeInstance($userAgent);
+
+        return $staticContainer->isLibrary();
     }
 
 }