+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
 
-# 0.1.0
+# 4.1.0
+- Exclude user agents from load testing services
+- Exclude user agents from server side tracking SDK by enabling an option
+
+# 4.0.0
 
 * Initial version